refactor: avoid package circular imports (deep imports, no barrel)
Made-with: Cursor
