+++ packages/open-ui-kit/src/components/actions-modal/components/comment-suggestions.tsx
@@ -8,7 +8,7 @@ import React from "react";
 import { GeneralSize } from "@/common";
 import { Stack } from "@mui/material";
 import { styles } from "../styles";
-import { Tag, TagBackgroundColorVariants } from "@/components";
+import { Tag, TagBackgroundColorVariants } from "@/components/tags/tag";
 
 type CommentSuggestionsProps = {
   commentSuggestions: string[];
